fix search screen
