+++ presentation/src/main/java/com/jusiCool/presentation/search/screen/SearchScreen.kt
@@ -1,14 +1,21 @@
 package com.jusiCool.presentation.search.screen
 
 import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.drawBehind
+import androidx.compose.ui.geometry.Offset
+import androidx.compose.ui.geometry.Size
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.navigation.NavController
@@ -18,13 +25,9 @@ import com.example.design_system.component.modifier.clickableSingle.clickableSin
 import com.example.design_system.icon_image.icon.LeftArrowIcon
 import com.example.design_system.theme.JDSTypography
 import com.example.design_system.theme.color.JDSColor
-import com.jusiCool.presentation.orderHistory.screen.OrderHistoryRoute
-import com.jusiCool.presentation.orderHistory.screen.OrderHistoryScreen
-import com.jusiCool.presentation.orderHistory.screen.tempMyStocksOrderHistoryData
-import com.jusiCool.presentation.orderHistory.screen.tempMyStocksOrderReservationData
 import com.jusiCool.presentation.search.component.PopularStocksSearch
 import com.jusiCool.presentation.search.component.PopularStocksSearchData
-import com.jusiCool.presentation.search.component.SearchTopBar
+import com.jusiCool.presentation.search.component.SearchTextField
 import kotlinx.collections.immutable.ImmutableList
 import kotlinx.collections.immutable.persistentListOf
 
@@ -75,11 +78,30 @@ internal fun SearchScreen(
             .fillMaxSize()
             .background(color = JDSColor.WHITE)
     ) {
-        SearchTopBar(
-            startIcon = { LeftArrowIcon(modifier = Modifier.clickableSingle { popUpBackStack() }) },
-            textState = stockTextState,
-            onTextChange = setStockTextState
-        )
+        Row(
+            horizontalArrangement = Arrangement.spacedBy(12.dp),
+            verticalAlignment = Alignment.CenterVertically,
+            modifier = modifier
+                .fillMaxWidth()
+                .drawBehind {
+                    drawRect(
+                        JDSColor.GRAY100,
+                        Offset(0f, size.height - 1.dp.toPx()),
+                        Size(size.width, 1.dp.toPx())
+                    )
+                }
+                .padding(
+                    horizontal = 24.dp,
+                    vertical = 16.dp
+                )
+        ) {
+            LeftArrowIcon( modifier = Modifier.clickableSingle { popUpBackStack() })
+
+            SearchTextField(
+                textState = stockTextState,
+                onTextChange = setStockTextState
+            )
+        }
 
         if(stockTextState.isEmpty()) {
             Text(
